# spec/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/socks/{id}":{"put":{"tags":["socks-controller"],"summary":"Updates existing socks by id","operationId":"updateSocks","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SocksDTO"}}},"required":true},"responses":{"200":{"description":"OK"},"400":{"description":"Bad request"}}}},"/api/socks/outcome":{"post":{"tags":["socks-controller"],"summary":"Decreases quantity of existing socks","operationId":"decreaseSocks","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SocksDTO"}}},"required":true},"responses":{"200":{"description":"OK"},"400":{"description":"Bad request"}}}},"/api/socks/income":{"post":{"tags":["socks-controller"],"summary":"Adds socks or increases quantity of existing ones","operationId":"addOrIncreaseSocks","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SocksDTO"}}},"required":true},"responses":{"200":{"description":"OK"},"400":{"description":"Bad request"}}}},"/api/socks/batch":{"post":{"tags":["socks-controller"],"summary":"Adds or increases the quantity of socks from a CSV file","operationId":"addOrIncreaseSocksFromCSV","requestBody":{"content":{"application/json":{"schema":{"required":["file"],"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK"},"400":{"description":"Bad request"}}}},"/api/socks":{"get":{"tags":["socks-controller"],"summary":"Returns quantity of socks based on provided filters","operationId":"getSocksQuantityWithFilters","parameters":[{"name":"color","in":"query","required":false,"schema":{"type":"string"}},{"name":"comparisonOperator","in":"query","required":false,"schema":{"type":"string"}},{"name":"cottonPercentage","in":"query","required":false,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"integer","format":"int32"}}}},"400":{"description":"Bad request","content":{"*/*":{"schema":{"type":"integer","format":"int32"}}}}}}}},"components":{"schemas":{"SocksDTO":{"type":"object","properties":{"color":{"type":"string"},"cottonPercentage":{"type":"integer","format":"int32"},"quantity":{"type":"integer","format":"int32"}}}}}}